requirements.txt
Django==1.11.4
django-allauth==0.32.0
django-angular==0.8.3
django-debug-toolbar==1.6
django-debug-toolbar-request-history==0.0.3
django-debug-toolbar-template-profiler==1.0.1
django-debug-toolbar-template-timings==0.7
djangorestframework==3.5.3
requests==2.18.4
oauthlib==2.0.2
requests-oauthlib==0.8.0
python-openid==2.2.5
pytz==2017.2
mysqlclient==1.3.10